Added filters

+++ ASPNetCoreMastersTodoList.Api/Controllers/ItemsController.cs
@@ -1,10 +1,12 @@
 ﻿using ASPNetCoreMastersTodoList.Api.BindingModels;
+using ASPNetCoreMastersTodoList.Api.Filters;
 using Microsoft.AspNetCore.Mvc;
 using Repositories;
 using Services.DTO;
 using Services.Interfaces;
 using System;
 using System.Collections.Generic;
+using System.Net;
 
 namespace ASPNetCoreMastersTodoList.Controllers
 {
@@ -26,8 +28,13 @@ namespace ASPNetCoreMastersTodoList.Controllers
         }
 
         [HttpGet("{itemId}")]
+        [ItemValidationFilterAttribute]
         public IActionResult Get(int itemId)
         {
+            if (Response.StatusCode.Equals((int)HttpStatusCode.NotFound))
+            {
+                return NotFound();
+            }
             return Ok(_itemService.GetById(itemId));
         }
 
@@ -49,15 +56,25 @@ namespace ASPNetCoreMastersTodoList.Controllers
         }
 
         [HttpPut()]
+        [ItemValidationFilterAttribute]
         public IActionResult Put(ItemDto item)
         {
+            if (Response.StatusCode.Equals((int)HttpStatusCode.NotFound))
+            {
+                return NotFound();
+            }
             _itemService.Update(item);
             return Ok();
         }
 
         [HttpDelete("{itemId}")]
+        [ItemValidationFilterAttribute]
         public IActionResult Delete(int itemId)
         {
+            if (Response.StatusCode.Equals((int)HttpStatusCode.NotFound))
+            {
+                return NotFound();
+            }
             _itemService.Delete(itemId);
             return Ok();
         }
